Correction of some warnings

+++ planning/include/planning/modules/SensorModel.h
@@ -28,14 +28,14 @@ public:
     int p_resolution_y_;
     double p_sampling_time_;  // sample camera poses from segment, use 0 for last
                                 // only
-    SensorModel(double ray_length, double fov_x, double fov_y, double resolution_x, double resolution_y, double sampling_time);
+    SensorModel(double ray_length, double fov_x, double fov_y, int resolution_x, int resolution_y, double sampling_time);
     SensorModel();
     ~SensorModel();
     void getDirectionVector_LIDAR(Eigen::Vector3d* result, double relative_x,
                                     double relative_y);
 };
 
-SensorModel::SensorModel(double ray_length, double fov_x, double fov_y, double resolution_x, double resolution_y, double sampling_time) : p_ray_length_{ray_length}, p_fov_x_{fov_x}, p_fov_y_{fov_y}, p_resolution_x_{resolution_x}, p_resolution_y_{resolution_y}, p_sampling_time_{sampling_time}
+SensorModel::SensorModel(double ray_length, double fov_x, double fov_y, int resolution_x, int resolution_y, double sampling_time) : p_ray_length_{ray_length}, p_fov_x_{fov_x}, p_fov_y_{fov_y}, p_resolution_x_{resolution_x}, p_resolution_y_{resolution_y}, p_sampling_time_{sampling_time}
 {
 }
